Made several changes pertaining to IdeaLab
